+++ fmri_autoreg/models/train_model.py
@@ -6,9 +6,9 @@ import pickle as pk
 import csv
 from math import ceil
 from sklearn.metrics import r2_score
-from src.data.load_data import load_params, load_data, make_input_labels
-from src.models.make_model import make_model
-from src.tools import check_path
+from fmri_autoreg.data.load_data import load_params, load_data, make_input_labels
+from fmri_autoreg.models.make_model import make_model
+from fmri_autoreg.tools import check_path
 
 
 def train(params, data, verbose=1):
